+++ azext_iot/tests/iothub/test_iot_ext_unit.py
@@ -1578,7 +1578,8 @@ class TestDeviceMethodInvoke:
         if methodbody:
             assert body["payload"] == json.loads(payload)
         else:
-            assert body.get("payload") is None
+            # We must ensure null is passed for payload.
+            assert body["payload"] is None
 
         assert body["responseTimeoutInSeconds"] == timeout
         assert body["connectTimeoutInSeconds"] == timeout
@@ -1658,7 +1659,8 @@ class TestDeviceModuleMethodInvoke:
         if methodbody:
             assert body["payload"] == json.loads(payload)
         else:
-            assert body.get("payload") is None
+            # We must ensure null is passed for payload.
+            assert body["payload"] is None
 
         assert body["responseTimeoutInSeconds"] == timeout
         assert body["connectTimeoutInSeconds"] == timeout